backend/requirements.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.4.0
django-celery-beat==2.9.0
django-celery-results==2.5.1
Pillow==10.4.0
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.7.0
openpyxl==3.1.5
reportlab==4.2.5
drf-spectacular==0.27.2
django-extensions==3.2.3
